fix: bim-icon svg image position & bim-label font-size/line height #56. (#67)
* fix: bim-icon scg image position & bim-label font-size/line height #56.

For this one #56 I fixed the initial bug for the svg image by adding display: flex; in the style of packages\core\src\components\Icon.ts.

I also fixed a related issue with regards to the bim-label size and line height with ref #55, by changing line-height from 1 rem to 1em, from packages\core\src\components\Label.ts style. We could also remove the property entirely and it would work.

To test it you can go to the Label html file change the font-size and the text and see that the icon will always be on the same place and the line-height now adjusts to the size.

Reviewed-by: Goulielmos Floros
Ref: #56

* fix: updates on examples

* update examples